:art: Create a helper to get Mundipagg admin menu.
