+++ system/library/mundipagg/src/Controller/Events.php
@@ -2,6 +2,7 @@
 namespace Mundipagg\Controller;
 
 use Mundipagg\Model\Order;
+use Mundipagg\Helper\AdminMenu as MundipaggHelperAdminMenu;
 
 require_once DIR_SYSTEM . 'library/mundipagg/vendor/autoload.php';
 
@@ -86,7 +87,8 @@ class Events
 
     public function columnLeftEntry($data)
     {
-        $mundipaggMenu = $this->getMundipaggMenu();
+        $mundipaggMenuHelper = new MundipaggHelperAdminMenu($this->openCart);
+        $mundipaggMenu = $mundipaggMenuHelper->getMenu();
 
         array_unshift($data['menus'], $mundipaggMenu);
 
@@ -97,43 +99,5 @@ class Events
         return $this->template;
     }
 
-    private function getMundipaggMenu()
-    {
-        $htmlLogo =
-            $this
-                ->openCart
-                ->load
-                ->view('extension/payment/mundipagg/menu/mundipagg');
 
-        $children[] = $this->getMenuChildren('Settings');
-        $children[] = $this->getMenuChildren('Subscriptions');
-        $children[] = $this->getMenuChildren('Plans');
-
-        $mundipaggMenu = [
-            'id'       => 'menu-mundipagg',
-            'name'	   => $htmlLogo,
-            'children' => $children
-        ];
-
-        return $mundipaggMenu;
-    }
-
-    private function getMenuChildren($name)
-    {
-        $path = 'extension/payment/mundipagg/' . strtolower($name);
-
-        return [
-            'name'  => $name,
-            'href'  => $this->getLink($path)
-        ];
-    }
-
-    private function getLink($path)
-    {
-        return $this->openCart->url->link(
-            $path,
-            'user_token=' . $this->openCart->session->data['user_token'],
-            true
-        );
-    }
 }
